+++ controllers/BookController.php
@@ -66,6 +66,10 @@ class BookController extends Controller
      */
     public function actionCreate()
     {
+	    if(Yii::$app->user->isGuest)
+	    {
+		    return $this->redirect(Yii::$app->homeUrl);
+	    }
         $model = new Book();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -87,6 +91,10 @@ class BookController extends Controller
      */
     public function actionUpdate($id)
     {
+	    if(Yii::$app->user->isGuest)
+	    {
+		    return $this->redirect(Yii::$app->homeUrl);
+	    }
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -108,6 +116,10 @@ class BookController extends Controller
      */
     public function actionDelete($id)
     {
+	    if(Yii::$app->user->isGuest)
+	    {
+		    return $this->redirect(Yii::$app->homeUrl);
+	    }
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
